# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medical Records API",
        "description": "API quan ly ho so y te",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/doctors": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Danh sách bác sĩ",
                "operationId": "a63fa5376d8d068fec1f154c81c31604",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Nguyễn"
                        }
                    },
                    {
                        "name": "specialty",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "cardiology"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách bác sĩ"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Tạo bác sĩ mới",
                "operationId": "92247d401ad549b11508c44d3288bdac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "personal_info.full_name": {
                                        "type": "string",
                                        "example": "Dr. Nguyễn Văn A"
                                    },
                                    "personal_info.birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-01-15"
                                    },
                                    "personal_info.gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "personal_info.phone": {
                                        "type": "string",
                                        "example": "0901234567"
                                    },
                                    "personal_info.email": {
                                        "type": "string",
                                        "example": "doctor@example.com"
                                    },
                                    "professional_info.specialty": {
                                        "type": "string",
                                        "example": "cardiology"
                                    },
                                    "professional_info.license_number": {
                                        "type": "string",
                                        "example": "DOC123456"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/v1/doctors/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Chi tiết bác sĩ",
                "operationId": "3921fa1a666c3d4e92bb5377c9ddd792",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin bác sĩ"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "put": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Cập nhật bác sĩ",
                "operationId": "1541cfbb09eb1767d80ab95009345166",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_rev": {
                                        "type": "string",
                                        "example": "1-abc123"
                                    },
                                    "personal_info.full_name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Xóa bác sĩ",
                "operationId": "0256468438b413ccd4c4db84945ae3a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rev",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1-abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "400": {
                        "description": "Thiếu rev parameter"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
...
    },
...
}